refactor(dataset): update imports across the codebase
